+++ test/e2e/dataset-get-sort-user-fields.spec.js
@@ -36,7 +36,7 @@ const mockDatasetsForSorting = async () => {
     ]);
 };
 
-describe('GET datasets sorted by user fields', () => {
+describe('Get datasets sorted by user fields', () => {
     before(async () => {
         if (process.env.NODE_ENV !== 'test') {
             throw Error(`Running the test suite with NODE_ENV ${process.env.NODE_ENV} may result in permanent data loss. Please use NODE_ENV=test.`);